backend/requirements.txt
Flask
Flask-CORS
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib
seaborn
openai
requests
beautifulsoup4
python-dateutil
statsmodels
prophet
xgboost
tensorflow
jieba
wordcloud
Pillow
tslearn
fastapi
uvicorn